Remove TransactionFiatTiles abstraction Adhere to a new convention of not arbitrarily grouping elements together that are not really a visual grouping
