+++ src/components/scenes/TransactionDetailsScene.tsx
@@ -1,4 +1,4 @@
-import { eq } from 'biggystring'
+import { abs, eq } from 'biggystring'
 import { EdgeCurrencyWallet, EdgeMetadata, EdgeTransaction } from 'edge-core-js'
 import * as React from 'react'
 import { TouchableWithoutFeedback, View } from 'react-native'
@@ -9,9 +9,18 @@ import { sprintf } from 'sprintf-js'
 import { formatCategory, getTxActionDisplayInfo, joinCategory, splitCategory } from '../../actions/CategoriesActions'
 import { playSendSound } from '../../actions/SoundActions'
 import { TX_ACTION_LABEL_MAP } from '../../constants/txActionConstants'
+import { getSymbolFromCurrency } from '../../constants/WalletAndCurrencyConstants'
 import { useContactThumbnail } from '../../hooks/redux/useContactThumbnail'
+import { displayFiatAmount } from '../../hooks/useFiatText'
+import { useHandler } from '../../hooks/useHandler'
+import { useHistoricalRate } from '../../hooks/useHistoricalRate'
+import { useWatch } from '../../hooks/useWatch'
 import { lstrings } from '../../locales/strings'
+import { getExchangeDenomination } from '../../selectors/DenominationSelectors'
+import { convertCurrencyFromExchangeRates } from '../../selectors/WalletSelectors'
+import { useSelector } from '../../types/reactRedux'
 import { EdgeSceneProps } from '../../types/routerTypes'
+import { convertNativeToExchange } from '../../util/utils'
 import { getMemoTitle } from '../../util/validateMemos'
 import { NotificationSceneWrapper } from '../common/SceneWrapper'
 import { withWallet } from '../hoc/withWallet'
@@ -26,7 +35,6 @@ import { EdgeText } from '../themed/EdgeText'
 import { MainButton } from '../themed/MainButton'
 import { SwapDetailsTiles } from '../tiles/SwapDetailsTiles'
 import { TransactionCryptoAmountTile } from '../tiles/TransactionCryptoAmountTile'
-import { TransactionFiatTiles } from '../tiles/TransactionFiatTiles'
 import { CardUi4 } from '../ui4/CardUi4'
 import { RowUi4 } from '../ui4/RowUi4'
 
@@ -43,7 +51,8 @@ export interface TransactionDetailsParams {
 const TransactionDetailsComponent = (props: Props) => {
   const { navigation, route, wallet } = props
   const { edgeTransaction: transaction, tokenId, walletId } = route.params
-  const { metadata, action } = transaction
+  const { metadata = {}, action, nativeAmount, date, currencyCode } = transaction
+  const { currencyInfo } = wallet
 
   const theme = useTheme()
   const styles = getStyles(theme)
@@ -83,6 +92,64 @@ const TransactionDetailsComponent = (props: Props) => {
   const [acceleratedTx, setAcceleratedTx] = React.useState<null | EdgeTransaction>(null)
 
   const { name = '' } = localMetadata
+
+  // #region Crypto Fiat Rows
+
+  // Look up wallet stuff:
+  const isoFiatCurrencyCode = useWatch(wallet, 'fiatCurrencyCode')
+  const fiatCurrencyCode = isoFiatCurrencyCode.replace('iso:', '')
+  const fiatSymbol = getSymbolFromCurrency(fiatCurrencyCode)
+
+  // Look up transaction stuff:
+  const absoluteAmount = abs(nativeAmount)
+
+  // Look up the current price:
+  const exchangeDenom = useSelector(state => getExchangeDenomination(state, currencyInfo.pluginId, currencyCode))
+  const absExchangeAmount = convertNativeToExchange(exchangeDenom.multiplier)(absoluteAmount)
+  const currentFiat = useSelector(state =>
+    parseFloat(convertCurrencyFromExchangeRates(state.exchangeRates, currencyCode, isoFiatCurrencyCode, absExchangeAmount))
+  )
+
+  // Look up the historical price:
+  const isoDate = new Date(date * 1000).toISOString()
+  const historicRate = useHistoricalRate(`${currencyCode}_${isoFiatCurrencyCode}`, isoDate)
+  const historicFiat = historicRate * Number(absExchangeAmount)
+
+  // Figure out which amount to show:
+  const displayFiat = metadata.amountFiat == null || metadata.amountFiat === 0 ? historicFiat : Math.abs(metadata.amountFiat)
+
+  // Percent difference:
+  const percentChange = displayFiat === 0 ? 0 : (100 * (currentFiat - displayFiat)) / displayFiat
+
+  // Convert to text:
+  const currentFiatText = displayFiatAmount(currentFiat)
+  const displayFiatText = displayFiatAmount(displayFiat)
+  const percentText = abs(percentChange.toFixed(2))
+
+  const handleEdit = useHandler(() => {
+    Airship.show<string | undefined>(bridge => (
+      <TextInputModal
+        bridge={bridge}
+        initialValue={displayFiatText}
+        inputLabel={fiatCurrencyCode}
+        returnKeyType="done"
+        keyboardType="numeric"
+        submitLabel={lstrings.string_save}
+        title={sprintf(lstrings.transaction_details_amount_in_fiat, fiatCurrencyCode)}
+      />
+    ))
+      .then(async inputText => {
+        if (inputText == null) return
+        const amountFiat = parseFloat(inputText.replace(',', '.'))
+
+        // Check for NaN, Infinity, and 0:
+        if (amountFiat === 0 || JSON.stringify(amountFiat) === 'null') return
+        await onSaveTxDetails({ amountFiat })
+      })
+      .catch(showError)
+  })
+
+  // #endregion Crypto Fiat Rows
 
   React.useEffect(() => {
     // Try accelerating transaction to check if transaction can be accelerated
@@ -210,7 +277,21 @@ const TransactionDetailsComponent = (props: Props) => {
 
       <CardUi4>
         <TransactionCryptoAmountTile transaction={transaction} wallet={wallet} />
-        <TransactionFiatTiles transaction={transaction} wallet={wallet} onMetadataEdit={onSaveTxDetails} />
+        <RowUi4 type="editable" title={sprintf(lstrings.transaction_details_amount_in_fiat, fiatCurrencyCode)} onPress={handleEdit}>
+          <View style={styles.tileRow}>
+            <EdgeText>{fiatSymbol + ' '}</EdgeText>
+            <EdgeText>{displayFiatText}</EdgeText>
+          </View>
+        </RowUi4>
+        <RowUi4 type="default" title={lstrings.transaction_details_amount_current_price}>
+          <View style={styles.tileRow}>
+            <EdgeText>{fiatSymbol + ' '}</EdgeText>
+            <EdgeText style={styles.tileTextPrice}>{currentFiatText}</EdgeText>
+            <EdgeText style={percentChange >= 0 ? styles.tileTextPriceChangeUp : styles.tileTextPriceChangeDown}>
+              {(percentChange >= 0 ? percentText : `- ${percentText}`) + '%'}
+            </EdgeText>
+          </View>
+        </RowUi4>
         {acceleratedTx == null ? null : (
           <RowUi4 type="touchable" title={lstrings.transaction_details_advance_details_accelerate} onPress={openAccelerateModel} />
         )}
@@ -260,6 +341,19 @@ const getStyles = cacheStyles((theme: Theme) => ({
   tileCategory: {
     marginVertical: theme.rem(0.25),
     color: theme.primaryText
+  },
+  tileTextPrice: {
+    flex: 1,
+    color: theme.primaryText,
+    fontSize: theme.rem(1)
+  },
+  tileTextPriceChangeUp: {
+    color: theme.positiveText,
+    fontSize: theme.rem(1)
+  },
+  tileTextPriceChangeDown: {
+    color: theme.negativeText,
+    fontSize: theme.rem(1)
   },
   textAdvancedTransaction: {
     color: theme.textLink,
